+++ plugins/__init__.py
@@ -1,0 +1,1 @@
+"""EEA bundled and third-party plugin packages."""