+++ plugins/builtin/__init__.py
@@ -1,0 +1,1 @@
+"""Official EEA bundled plugins."""